TopDownShooter: Rework input and add input tests.

+++ Executable/TopDownShooter/Model/Shared/Constants.hpp
@@ -3,4 +3,4 @@
 #include <cstdint>
 #include <numeric>
 
-constexpr std::uint64_t INVALID_OBJECT_ID = std::numeric_limits<std::uint64_t>::max ();
+constexpr std::uintptr_t INVALID_OBJECT_ID = std::numeric_limits<std::uintptr_t>::max ();
